requirements.txt
flask==3.0.0
groq==0.4.2
tavily-python==0.3.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
python-docx==1.1.0
markdown==3.5.1
flask-session==0.5.0
redis==5.0.1
python-dotenv==1.0.0
werkzeug==3.0.1